backend/requirements.txt
amqp==5.2.0
anyio==4.6.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
cssselect2==0.7.0
daphne==4.1.2
diff-match-patch==20241021
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-import-export==4.2.0
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
lxml==5.3.0
msgpack==1.1.0
oscrypto==1.3.0
pillow==10.4.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==5.0.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
redis==5.1.1
reportlab==4.2.5
requests==2.32.3
rsa==4.9
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
svglib==1.5.1
tablib==3.7.0
tinycss2==1.3.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
uvicorn==0.31.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
xhtml2pdf==0.2.16
zope.interface==7.0.3
